separate BaseViewHolder from Base class
